server/requirements.txt
django==2.1.5
djangorestframework==3.9.0
django-cors-headers==2.4.0
markdown==3.0.1
django-filter==2.0.0
pymysql==0.9.2
mysqlclient==1.3.13
pycrypto==2.6.1
celery==4.2.1
redis==3.2.0
requests==2.21.0